Spezzato il file in due
